Refactor errors.
