+++ frontend/src/actions/courses.ts
@@ -1,68 +1,96 @@
 'use server'
 
-import {redirect} from 'next/navigation'
+import { redirect } from 'next/navigation'
 
-import {CoursePublic, CoursesService} from '@/client'
-import {zCourseCreate} from '@/client/zod.gen'
+import { CoursePublic, CoursesService, CourseWithDocuments } from '@/client'
+import { zCourseCreate } from '@/client/zod.gen'
 
-import {handleError} from './handleErrors'
-import {IState} from '@/types/common'
+import { mapApiError } from '@/lib/mapApiError'
+import { Result } from '@/lib/result'
 
-export async function getCourses(): Promise<CoursePublic[] | IState> {
+/**
+ * Fetch all courses.
+ * Returns a Result<CoursePublic[]> that must be checked with `res.ok`.
+ */
+export async function getCourses(): Promise<Result<CoursePublic[]>> {
   try {
     const response = await CoursesService.getApiV1Courses({
       responseValidator: async () => {},
     })
-    return response.data?.data ?? []
-  } catch (error) {
+
     return {
-      message: handleError(error),
-      success: false,
-      courses: [],
+      ok: true,
+      data: response.data?.data ?? [],
+    }
+  } catch (err) {
+    return {
+      ok: false,
+      error: mapApiError(err),
     }
   }
 }
 
-export async function getCourse(id: string): Promise<CoursePublic | IState> {
+/**
+ * Fetch a single course by ID.
+ * Returns a Result<CoursePublic>.
+ */
+export async function getCourse(id: string): Promise<Result<CourseWithDocuments>> {
   try {
     const response = await CoursesService.getApiV1CoursesById({
-      path: {id},
+      path: { id },
       responseValidator: async () => {},
     })
-    return response.data
-  } catch (error) {
+
     return {
-      message: handleError(error),
-      success: false,
+      ok: true,
+      data: response.data,
+    }
+  } catch (err) {
+    return {
+      ok: false,
+      error: mapApiError(err),
     }
   }
 }
 
-export async function createCourse(_state: unknown, formData: FormData) {
-  let course: CoursePublic | IState
+/**
+ * Create a course from FormData.
+ * Returns a Result<CoursePublic> that must be checked with `res.ok`.
+ * If successful, it redirects to the dashboard page.
+ */
+export async function createCourse(_state: unknown, formData: FormData): Promise<Result<CoursePublic>> {
   try {
     const payload = zCourseCreate.safeParse({
       name: formData.get('name'),
       description: formData.get('description'),
     })
+
     if (!payload.success) {
       return {
-        success: false,
-        message: 'Validation failed',
-        errors: payload.error.flatten().fieldErrors,
+        ok: false,
+        error: {
+          code: 'VALIDATION',
+          message: 'Validation failed',
+          details: payload.error.flatten().fieldErrors,
+        },
       }
     }
 
     const response = await CoursesService.postApiV1Courses({
       body: payload.data,
     })
-    course = response.data
-  } catch (error) {
+
+    // Redirect after creation — safe to redirect only on success
+    redirect(`/dashboard/courses/create?courseId=${response.data.id}`)
+
     return {
-      message: handleError(error),
-      success: false,
+      ok: true,
+      data: response.data,
+    }
+  } catch (err) {
+    return {
+      ok: false,
+      error: mapApiError(err),
     }
   }
-
-  redirect(`/dashboard/courses/create?courseId=${course?.id}`)
 }
